requirements.txt
unidiff
requests
detect_secrets
sqlitedict
beautifulsoup4
python-dateutil
textract